Added API requests, disclaimers, edited enemy data
An api key for the JailBase.com api is necessary to make requests.
This key should be placed in 'data/key.txt'

The TOS only allows archival for up to 24 hours, so I will not be adding
a method to load a saved list.

The gathered list is not currently used, this commit only includes the
gathering.

Only name, picture, charges, and an external link to the data(required
by TOS) are included.

Not all exceptions are handled at the moment.

Will be adding additional disclaimers to project in accordance with
JailBase TOS.